cpufreq: stats: update uid data in acct_update_power
This eliminates the need to maintain separate alive_* and dead_*
arrays for each uid, as well as the need to loop over every task and
repeatedly take locks when printing out per-uid stats.

It also should facilitate replacing the use of single_open with seq
iterator because readers no longer need to take uid_lock or modify
uid_hashtable.

Test: Read /proc/uid_time_in_state, /proc/uid_cpupower/time_in_state
and /proc/uid/<uid>/time_in_state and check that values are
reasonable.
Test: Reading /proc/uid_cpupower/time_in_state takes <1ms
Bug: 71718257
Bug: 73328000
Change-Id: Ic5fbb7d19d1c1e2c1c57e1ee6622e3b144acc9d4
Signed-off-by: Connor O'Brien <connoro@google.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>